requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
pyyaml>=6.0
faker>=18.0.0
click>=8.0.0
python-dotenv>=1.0.0
